requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.0.0
altgraph==0.17.4
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
entrypoints==0.4
faiss-cpu==1.7.4
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.0.184
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numexpr==2.10.0
numpy==1.26.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==24.0
pandas==2.2.2
pdf2image==1.17.0
pdfminer==20191125
pillow==10.3.0
protobuf==3.20.3
pyarrow==16.1.0
pycryptodome==3.17
pydantic==1.10.15
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
semver==3.0.2
six==1.16.0
smmap==5.0.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
streamlit==1.18.1
tenacity==8.3.0
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.28.3
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.18.2
